% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_head_na}
\alias{fill_head_na}
\title{fill NA which is "head" of a data.frame to other value.}
\usage{
fill_head_na(data, cols, value)
}
\arguments{
\item{data}{a data.frame}

\item{cols}{<tidy-select> a column or columns}

\item{value}{value which NA are replaced with}
}
\value{
a data.frame
}
\description{
fill NA which is "head" of a data.frame to other value.
}
\examples{
\dontrun{
  df <- data.frame(aa=c(NA,NA,1,2,NA),ab=c(NA,2,3,NA,4),ac=c(1,2,NA,3,NA))
  fill_head_na(df,aa,0)
  fill_head_na(df,c(aa,ab),0)
  fill_head_na(df,everything(),0)
}
}
