% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_na.R
\name{interpolate_na}
\alias{interpolate_na}
\title{interpolate NA, only at the "head" and "tail" of a data.frame.}
\usage{
interpolate_na(data, cols, fn, ...)
}
\arguments{
\item{data}{a data.frame}

\item{cols}{<tidy-select> a column or columns}

\item{fn}{a interpolate function}

\item{...}{passing to fn}
}
\value{
a data.frame
}
\description{
interpolate NA, only at the "head" and "tail" of a data.frame.
}
\examples{
\dontrun{
  df <- data.frame(aa=c(NA,NA,1,2,NA),ab=c(NA,2,3,NA,4),ac=c(1,2,NA,3,NA))
  interpolate_na(df,aa,zoo::na.approx)
  interpolate_na(df,c(aa,ab),zoo::na.approx)
  interpolate_na(df,everything(),zoo::na.spline)
}
}
