% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sub_bytes.R
\name{str_sub_bytes}
\alias{str_sub_bytes}
\title{Extract substrings from a character vector used by bytes.}
\usage{
str_sub_bytes(str, start, end, bytes_type = c("Shift-JIS", "Unicode"))
}
\arguments{
\item{str}{input vector}

\item{start}{numeric}

\item{end}{numeric}

\item{bytes_type}{"Shift-JIS" or "Unicode"
"Shift-JIS": Every characters are counted by a rule of Shift-JIS, that is,
ASCII characters are counted as 1 byte and non-ASCII characters are counted as 2 bytes.
"Unicode": Every characters are counted by a rule of Unicode, that is,
ASCII characters are counted as 1 bytes and non-ASCII characters are counted as 2 bytes to 4 bytes.}
}
\value{
a character of character vector
}
\description{
Extract substrings from a character vector used by bytes.
}
\examples{
\dontrun{
  str_sub_bytes("abcdefg",3,6,"Unicode")
}
}
